+++ lib/api/base.dart
@@ -5,11 +5,13 @@ const kProtobufContentType = 'application/x-protobuf';
 enum HttpMethod { post, get, probuf }
 
 const kMultipartFormDataHeader = 'multipart/form-data';
+
 typedef MyFormData = dio.FormData;
 typedef MyMultipartFile = dio.MultipartFile;
 typedef MyInterceptorWrapper = dio.InterceptorsWrapper;
 typedef ParseObject = BaseModel Function(Map<String, dynamic> originMap);
 typedef ToastWrapper = FlutterSmartDialog;
+
 @Doc(message: "简单toast弹窗")
 void toast(String msg) {
   SmartDialog.dismiss(status: SmartStatus.toast);
@@ -33,8 +35,7 @@ mixin BasePagedApiMixin on BaseApi {
 abstract class BaseApi {
   static bool showLog = false;
   static late String _host;
-  static dio.BaseOptions options =
-      dio.BaseOptions(connectTimeout: const Duration(milliseconds: 30000),receiveTimeout: const Duration(seconds: 5));
+  static dio.BaseOptions options = dio.BaseOptions(connectTimeout: const Duration(milliseconds: 30000), receiveTimeout: const Duration(seconds: 5));
 
   static set host(String h) => _host = h;
 
@@ -58,20 +59,11 @@ abstract class BaseApi {
       }
       final d = getDio();
       interceptions.add(ErrorInterceptor());
-      d.interceptors.addAll(
-          options.interceptorCall?.call(interceptions) ?? interceptions);
-      final contentTypeStr = options.contentType.isNotEmpty
-          ? options.contentType
-          : (httpMethod == HttpMethod.post ? io.ContentType.json.value : null);
-      final bodyParams =
-          formData.files.isNotEmpty ? formData : (options.data ?? params);
-
-      final queryParameters = httpMethod == HttpMethod.post
-          ? null
-          : (options.nullParams == true ? null : options.data ?? params);
-      final contentTypeString = httpMethod == HttpMethod.probuf
-          ? kProtobufContentType
-          : contentTypeStr;
+      d.interceptors.addAll(options.interceptorCall?.call(interceptions) ?? interceptions);
+      final contentTypeStr = options.contentType ?? (httpMethod == HttpMethod.post ? io.ContentType.json.value : options.contentType);
+      final bodyParams = formData.files.isNotEmpty ? formData : (options.data ?? params);
+      final queryParameters = httpMethod == HttpMethod.post ? null : (options.nullParams == true ? null : options.data ?? params);
+      final contentTypeString = httpMethod == HttpMethod.probuf ? kProtobufContentType : contentTypeStr;
       final finalUrl = options.isFullUrl ? url : (_host + url);
       await options.dioStart?.call(d, finalUrl);
       printLog("url---$finalUrl");
@@ -79,13 +71,14 @@ abstract class BaseApi {
       final response = await d.request(
         finalUrl,
         options: dio.Options(
-            method: method,
-            contentType: contentTypeString,
-            headers: options.headers,
-            responseType: options.responseType,
-            requestEncoder: options.requestEncoder,),
-        queryParameters: queryParameters,
-        data: bodyParams,
+          method: method,
+          contentType: contentTypeString,
+          headers: options.headers,
+          responseType: options.responseType,
+          requestEncoder: options.requestEncoder,
+        ),
+        queryParameters: httpMethod == HttpMethod.get ? queryParameters : null,
+        data: httpMethod == HttpMethod.get ? null : bodyParams,
       );
       if (options.showDefaultLoading) {
         closeLoading();
@@ -102,15 +95,12 @@ abstract class BaseApi {
           } catch (e) {
             kLogErr(data.runtimeType);
             kLogErr(data);
-            throw AppException.appError(
-                code: 10003, msg: "Unable to process server data", data: data);
+            throw AppException.appError(code: 10003, msg: "Unable to process server data", data: data);
           }
         }
         return data;
       } else {
-        throw AppException(
-            code: response.statusCode ?? 10004,
-            message: response.statusMessage ?? "ERROR");
+        throw AppException(code: response.statusCode ?? 10004, message: response.statusMessage ?? "ERROR");
       }
     } on dio.DioError catch (e) {
       if (options.showDefaultLoading) {
@@ -178,8 +168,7 @@ class Doc {
 
 ///wrapjson类型的接口封装
 abstract class AppCoreApi extends BaseApi {
-  AppCoreApi(String url, {HttpMethod? httpMethod, List<dio.Interceptor>? ints})
-      : super(url, httpMethod: httpMethod ?? HttpMethod.get) {
+  AppCoreApi(String url, {HttpMethod? httpMethod, List<dio.Interceptor>? ints}) : super(url, httpMethod: httpMethod ?? HttpMethod.get) {
     if (ints?.isNotEmpty == true) {
       interceptions.addAll(ints!);
     }