+++ lib/ext/context.dart
@@ -47,18 +47,17 @@ extension ContextExt on BuildContext {
 
   void pop() => Navigator.pop(this);
 
-  Future<void> showSimpleDialog(String tip,{String? title,String? cancelText}) async {
+  Future<void> showSimpleDialog(String tip, {String? title, String? cancelText}) async {
     await showCupertinoDialog(
-        context: this,
-        builder: (c) {
-          return CupertinoAlertDialog(
-            title:  Text(title ?? "提示"),
-            content: Text(tip),
-            actions: [
-              CupertinoDialogAction(onPressed: pop, child: Text(cancelText??'Cancel'))
-            ],
-          );
-        },);
+      context: this,
+      builder: (c) {
+        return CupertinoAlertDialog(
+          title: Text(title ?? "提示"),
+          content: Text(tip),
+          actions: [CupertinoDialogAction(onPressed: pop, child: Text(cancelText ?? 'Cancel'))],
+        );
+      },
+    );
   }
 
   Future<void> bottomSheet<T>({required Widget child}) async {
@@ -66,6 +65,17 @@ extension ContextExt on BuildContext {
   }
 
   Future<bool> askOk(AskOkDialogParams params) async {
-    return AskOkDialog.show(this,params);
+    return AskOkDialog.show(this, params);
+  }
+
+  Future<void> simpleAskOk(String content, VoidCallback trueCall, [AskOkDialogParams Function(AskOkDialogParams old)? paramsHandle]) async {
+    var params = AskOkDialogParams(contentText: content);
+    if (paramsHandle != null) {
+      params = paramsHandle.call(params);
+    }
+    final isOk = await askOk(params);
+    if(isOk){
+      trueCall.call();
+    }
   }
 }